Development

+++ src/de/k7bot/music/utilities/MusicUtil.java
@@ -1,16 +1,30 @@
 package de.k7bot.music.utilities;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.concurrent.TimeUnit;
 
+import org.apache.hc.core5.http.ParseException;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClients;
+import org.apache.http.util.EntityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
+import com.sedmelluq.discord.lavaplayer.source.youtube.YoutubeAudioTrack;
+import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 
 import de.k7bot.Klassenserver7bbot;
 import de.k7bot.music.lavaplayer.MusicController;
+import de.k7bot.music.utilities.spotify.SpotifyAudioTrack;
 import de.k7bot.sql.LiteSQL;
 import de.k7bot.util.GenericMessageSendHandler;
 import net.dv8tion.jda.api.EmbedBuilder;
@@ -22,6 +36,10 @@ import net.dv8tion.jda.api.entities.channel.middleman.AudioChannel;
 import net.dv8tion.jda.api.interactions.InteractionHook;
 import net.dv8tion.jda.api.managers.AudioManager;
 import net.dv8tion.jda.api.utils.FileUpload;
+import se.michaelthelin.spotify.SpotifyApi;
+import se.michaelthelin.spotify.exceptions.SpotifyWebApiException;
+import se.michaelthelin.spotify.model_objects.specification.Image;
+import se.michaelthelin.spotify.model_objects.specification.Track;
 
 /**
  *
@@ -38,9 +56,9 @@ public class MusicUtil {
 	 */
 	public static void updateChannel(TextChannel channel) {
 
-		ResultSet set = LiteSQL.onQuery("SELECT * FROM musicutil WHERE guildId = ?;", channel.getGuild().getIdLong());
+		try (ResultSet set = LiteSQL.onQuery("SELECT * FROM musicutil WHERE guildId = ?;",
+				channel.getGuild().getIdLong())) {
 
-		try {
 			if (set.next()) {
 				LiteSQL.onUpdate("UPDATE musicutil SET channelId = ? WHERE guildId = ?;", channel.getIdLong(),
 						channel.getGuild().getIdLong());
@@ -69,23 +87,38 @@ public class MusicUtil {
 	 * @param guildid
 	 * @param builder
 	 */
-	public static void sendEmbed(long guildid, EmbedBuilder builder) {
-		ResultSet set = LiteSQL.onQuery("SELECT channelId FROM musicutil WHERE guildId = ?", guildid);
+	public static void sendIconEmbed(long guildid, EmbedBuilder builder, AudioTrack track) {
 
-		try {
-			if (set.next()) {
-				long channelid = set.getLong("channelId");
-				Guild guild;
-				if ((guild = Klassenserver7bbot.getInstance().getShardManager().getGuildById(guildid)) != null) {
-					TextChannel channel;
-					if ((channel = guild.getTextChannelById(channelid)) != null) {
-						channel.sendMessageEmbeds(builder.build()).queue();
-					}
+		try (ResultSet set = LiteSQL.onQuery("SELECT channelId FROM musicutil WHERE guildId = ?", guildid)) {
+
+			if (!set.next()) {
+				return;
+			}
+
+			long channelid = set.getLong("channelId");
+			Guild guild;
+
+			if ((guild = Klassenserver7bbot.getInstance().getShardManager().getGuildById(guildid)) == null) {
+				return;
+			}
+
+			TextChannel channel;
+
+			if ((channel = guild.getTextChannelById(channelid)) != null) {
+
+				FileUpload up = setIcons(track);
+
+				if (up == null) {
+					channel.sendMessageEmbeds(builder.build()).queue();
+				} else {
+					builder.setImage("attachment://thumbnail.png");
+					channel.sendFiles(up).setEmbeds(builder.build()).queue();
+					up.close();
 				}
 
 			}
 
-		} catch (SQLException e) {
+		} catch (SQLException | IOException e) {
 			log.error(e.getMessage(), e);
 		}
 	}
@@ -95,20 +128,23 @@ public class MusicUtil {
 	 * @param guildid
 	 * @param builder
 	 */
-	public static void sendEmbed(long guildid, EmbedBuilder builder, FileUpload upload) {
-		ResultSet set = LiteSQL.onQuery("SELECT channelId FROM musicutil WHERE guildId = ?", guildid);
+	public static void sendEmbed(long guildid, EmbedBuilder builder) {
 
-		try {
-			if (set.next()) {
-				long channelid = set.getLong("channelId");
-				Guild guild;
-				if ((guild = Klassenserver7bbot.getInstance().getShardManager().getGuildById(guildid)) != null) {
-					TextChannel channel;
-					if ((channel = guild.getTextChannelById(channelid)) != null) {
-						builder.setImage("attachment://thumbnail.png");
-						channel.sendFiles(upload).setEmbeds(builder.build()).queue();
-					}
-				}
+		try (ResultSet set = LiteSQL.onQuery("SELECT channelId FROM musicutil WHERE guildId = ?", guildid)) {
+
+			if (!set.next()) {
+				return;
+			}
+			long channelid = set.getLong("channelId");
+			Guild guild;
+
+			if ((guild = Klassenserver7bbot.getInstance().getShardManager().getGuildById(guildid)) == null) {
+				return;
+			}
+
+			TextChannel channel;
+			if ((channel = guild.getTextChannelById(channelid)) != null) {
+				channel.sendMessageEmbeds(builder.build()).queue();
 
 			}
 
@@ -164,9 +200,9 @@ public class MusicUtil {
 
 		if ((ac.getIdLong() == audioman.getConnectedChannel().getIdLong())) {
 			return 1;
-		} else {
-			return 0;
 		}
+
+		return 0;
 	}
 
 	/**
@@ -183,7 +219,7 @@ public class MusicUtil {
 			return false;
 		}
 
-		return checkChannelConnection(sendHandler, m);
+		return checkChannelConnection(sendHandler, m, true);
 	}
 
 	/**
@@ -200,10 +236,11 @@ public class MusicUtil {
 			return false;
 		}
 
-		return checkChannelConnection(sendHandler, m);
+		return checkChannelConnection(sendHandler, m, false);
 	}
 
-	protected static boolean checkChannelConnection(GenericMessageSendHandler sendHandler, Member m) {
+	protected static boolean checkChannelConnection(GenericMessageSendHandler sendHandler, Member m,
+			boolean checkBeforePlay) {
 
 		AudioChannel vc = MusicUtil.getMembVcConnection(m);
 
@@ -214,6 +251,10 @@ public class MusicUtil {
 			return false;
 		}
 		case 2: {
+
+			if (checkBeforePlay) {
+				return true;
+			}
 
 			sendHandler.sendMessage("The Bot isn't playing a Song use -p [Song] to play one" + m.getAsMention())
 					.complete().delete().queueAfter(10L, TimeUnit.SECONDS);
@@ -277,5 +318,84 @@ public class MusicUtil {
 
 		return true;
 
+	}
+
+	private static FileUpload setIcons(AudioTrack track) {
+
+		if (track instanceof YoutubeAudioTrack) {
+			return loadYTIcon(track.getIdentifier());
+		}
+
+		if (track instanceof SpotifyAudioTrack) {
+			return loadSpotifyIcon(track.getIdentifier());
+		}
+		return null;
+
+	}
+
+	@SuppressWarnings("resource")
+	private static FileUpload loadSpotifyIcon(String songid) {
+
+		final HttpGet httpget = new HttpGet("https://open.spotify.com/get_access_token");
+
+		try (final CloseableHttpClient client = HttpClients.createSystem();
+				final CloseableHttpResponse response = client.execute(httpget)) {
+
+			if (response.getStatusLine().getStatusCode() != 200) {
+				return null;
+			}
+
+			JsonObject resp = JsonParser.parseString(EntityUtils.toString(response.getEntity())).getAsJsonObject();
+
+			String token = resp.get("accessToken").getAsString();
+
+			SpotifyApi api = SpotifyApi.builder().setAccessToken(token).build();
+
+			Track t = api.getTrack(songid).build().execute();
+
+			Image[] images = t.getAlbum().getImages();
+
+			Image img = images[0];
+
+			for (Image imgs : images) {
+				if (imgs.getHeight() > img.getHeight()) {
+					img = imgs;
+				}
+			}
+
+			InputStream file = (new URL(img.getUrl())).openStream();
+			return FileUpload.fromData(file, "thumbnail.jpg");
+
+		} catch (IOException | ParseException | SpotifyWebApiException e) {
+			log.error(e.getMessage(), e);
+			return null;
+		}
+
+	}
+
+	@SuppressWarnings("resource")
+	private static FileUpload loadYTIcon(String videoId) {
+
+		try {
+			InputStream file = (new URL("https://img.youtube.com/vi/" + videoId + "/maxresdefault.jpg")).openStream();
+
+			return FileUpload.fromData(file, "thumbnail.jpg");
+
+		} catch (IOException e) {
+
+			log.warn("No maxresdefault.jpg available");
+
+			try (InputStream file = (new URL("https://img.youtube.com/vi/" + videoId + "/hqdefault.jpg"))
+					.openStream()) {
+
+				return FileUpload.fromData(file, "thumbnail.jpg");
+
+			} catch (IOException e1) {
+				log.error(e.getMessage(), e);
+			}
+
+		}
+
+		return null;
 	}
 }
